+++ src/auth/interfaces/auth.controller.interface.ts
@@ -1,4 +1,5 @@
-import { FilteredPerson, FilteredPersonWithToken, SignInDTO, SignUpDTO } from '#auth/auth.types.js';
+import { FilteredPerson, FilteredPersonWithToken } from '#auth/types/filtered.types.js';
+import { SignInDTO, SignUpDTO } from '#auth/types/sign.dto.js';
 import { UserType } from '#types/common.types.js';
 import { Response } from 'express';
