+++ src/auth/types/auth.types.ts
@@ -1,0 +1,16 @@
+import { UserType } from '#types/common.types.js';
+
+export interface TokenPayload {
+  id: string;
+  type: UserType;
+}
+
+export class Tokens {
+  accessToken: string;
+  refreshToken?: string;
+}
+
+export const phoneNumberRegex =
+  /^(010\d{4}\d{4}|02\d{4}\d{4}|032\d{4}\d{4}|042\d{4}\d{4}|051\d{4}\d{4}|052\d{4}\d{4}|053\d{4}\d{4}|062\d{4}\d{4}|064\d{4}\d{4}|031\d{4}\d{4}|033\d{4}\d{4}|041\d{4}\d{4}|043\d{4}\d{4}|054\d{4}\d{4}|055\d{4}\d{4}|061\d{4}\d{4}|063\d{4}\d{4})$/;
+export const passwordRegex = /^(?=.*[a-zA-Z])(?=.*\d)(?=.*[!@#$%^&*])[a-zA-Z\d!@#$%^&*]{8,}$/;
+export const imageRegex = /(.*?)\.(jpg|jpeg|png|gif|bmp)$/;
