+++ src/auth/types/provider.dto.ts
@@ -1,0 +1,9 @@
+import { ProviderCreateDTO } from '#auth/types/provider.types.js';
+
+export interface GoogleCreateDTO extends ProviderCreateDTO {}
+
+export interface KakaoCreateDTO extends ProviderCreateDTO {
+  phoneNumber: string;
+}
+
+export interface NaverCreateDTO extends ProviderCreateDTO {}
