+++ src/drivers/interfaces/driver.service.interface.ts
@@ -1,6 +1,6 @@
+import { FilteredPersonalInfo } from '#auth/types/filtered.types.js';
 import { Driver, DriverPatchDTO } from '#drivers/driver.types.js';
 import { DriversFindOptions } from '#types/options.type.js';
-import { FilteredPersonalInfo } from '#types/personal.type.js';
 
 export interface IDriverService {
   findDrivers: (options: DriversFindOptions) => Promise<{ totalCount: number; list: FilteredPersonalInfo<Driver>[] }>;
