+++ src/swagger/swagger.service.ts
@@ -1,4 +1,5 @@
-import { FilteredDriverOutputDTO, FilteredUserOutputDTO } from '#auth/auth.types.js';
+import { FilteredDriverOutputDTO } from '#auth/types/filtered.driver.dto.js';
+import { FilteredUserOutputDTO } from '#auth/types/filtered.user.dto.js';
 import { swaggerConfig } from '#configs/swagger.config.js';
 import { QuestionEntity } from '#questions/question.types.js';
 import { INestApplication, Injectable } from '@nestjs/common';
