+++ src/utils/filterSensitiveData.ts
@@ -1,5 +1,5 @@
+import { FilteredPersonalInfo } from '#auth/types/filtered.types.js';
 import { Driver } from '#drivers/driver.types.js';
-import { FilteredPersonalInfo } from '#types/personal.type.js';
 import { User } from '#users/user.types.js';
 import { generateS3DownloadUrl } from '#utils/S3/generate-s3-download-url.js';
